Pull `step` up into `task` in remote db model
